:recycle: (MuscleGroupIcon) extract icon logic in generic component
