+++ src/components/Icons/MuscleGroupIcon/index.ts
