+++ src/pages/Home/Dashboard/components/MuscleGroupCard/index.ts
@@ -1,0 +1,1 @@
+export { MuscleGroupCard } from './MuscleGroupCard';
